enhance: add persistent DB triggers
This change allows for specifying a DSN that will be used to persist
triggers and the metadata to allow not triggering the world on startup.
For each kind, a resource version is tracked; any objects with a
resource version that has already been processed will not trigger other
objects to be processed. Additionally, objects that were already
processed will not be processed on startup.

If a type needs to have its objects reprocessed or its triggers rebuilt,
then the type can implement the triggers.Generationed interface to bump
the trigger generation. This will cause all objects to be reprocessed,
and the triggers that would trigger these objects to be rebuilt.

Signed-off-by: Donnie Adams <donnie@acorn.io>